working login and football api functionality
