+++ Backend/src/main/java/com/smartbet/CustomOidcUserService.java
@@ -5,6 +5,10 @@ import org.springframework.security.oauth2.client.oidc.userinfo.OidcUserService;
 import org.springframework.security.oauth2.core.OAuth2AuthenticationException;
 import org.springframework.security.oauth2.core.oidc.user.OidcUser;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import org.springframework.transaction.annotation.Propagation;
+
+import java.util.Optional;
 
 @Service
 public class CustomOidcUserService extends OidcUserService {
@@ -15,6 +19,7 @@ public class CustomOidcUserService extends OidcUserService {
     }
 
     @Override
+    @Transactional(propagation = Propagation.REQUIRES_NEW)
     public OidcUser loadUser(OidcUserRequest userRequest) throws OAuth2AuthenticationException {
         OidcUser oidcUser = super.loadUser(userRequest);
 
@@ -25,22 +30,24 @@ public class CustomOidcUserService extends OidcUserService {
         System.out.println("Google OIDC User - Subject: " + externalId);
         System.out.println("Google OIDC User attributes: " + oidcUser.getAttributes());
 
-        Optional<User> possibleUser = userService.findById(externalId);
-        User systemUser;
+        try {
+            Optional<User> possibleUser = userService.findById(externalId);
+            User systemUser;
 
-        if (possibleUser.isPresent()){
-            systemUser = possibleUser.get();
+            if (possibleUser.isPresent()) {
+                systemUser = possibleUser.get();
+                System.out.println("Found existing Google user: " + externalId);
+            } else {
+                systemUser = new User(externalId);
+                userService.createUser(systemUser);
+                System.out.println("Created new Google user: " + externalId);
+            }
+
+            return oidcUser;
+        } catch (Exception e) {
+            System.err.println("Error processing Google OIDC user: " + e.getMessage());
+            // Return the OIDC user even if database operations fail
+            return oidcUser;
         }
-        else{
-            systemUser = new User(externalId);
-            userService.createUser(systemUser);
-        }
-
-        // userService.findById(externalId).orElseGet(() -> {
-        //     User newUser = new User(externalId);
-        //     return userService.createUser(newUser);
-        // });
-
-        return oidcUser;
     }
 }
