+++ Backend/src/main/resources/application.yaml
@@ -2,11 +2,10 @@ spring:
   application:
     name: SmartBet
   datasource:
-    url: jdbc:postgresql://aws-0-ap-southeast-2.pooler.supabase.com:6543/postgres
+    url: jdbc:postgresql://aws-0-ap-southeast-2.pooler.supabase.com:5432/postgres?sslmode=require&preferQueryMode=simple
     username: postgres.kckimbkobioqyudqnzvg
     password: TheInvincibles@35
     driver-class-name: org.postgresql.Driver
-
     hikari:
       maximum-pool-size: 2
       minimum-idle: 0
@@ -14,9 +13,20 @@ spring:
       max-lifetime: 300000
       connection-timeout: 10000
       leak-detection-threshold: 60000
-  jpa:
-    hibernate:
-      ddl-auto: none                 # CRITICAL: Disable DDL auto
+
+#  jpa:
+#    hibernate:
+#      ddl-auto: none
+#    show-sql: false
+#    properties:
+#      hibernate:
+#        dialect: org.hibernate.dialect.PostgreSQLDialect
+#        format_sql: false
+#        jdbc:
+#          time_zone: UTC
+
+  main:
+    lazy-initialization: true
 
   security:
     oauth2:
@@ -26,8 +36,85 @@ spring:
             client-id: Ov23li8Lm4OC4u6A6LfN
             client-secret: 9ef36f34fa68771e9e8f67844b718237b3e8544a
             scope: read:user,user:email
-
           google:
             client-id: 462297550542-jlh81q4c3uho2afpl42kejgml7gkv7u3.apps.googleusercontent.com
             client-secret: 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
             scope: openid,profile,email
+
+  web:
+    resources:
+      static-locations: classpath:/static/
+
+  servlet:
+    multipart:
+      max-file-size: 10MB
+      max-request-size: 10MB
+
+server:
+  port: ${PORT:8080}
+  servlet:
+    context-path: /
+
+logging:
+  level:
+    org.hibernate: WARN
+    org.springframework.web: INFO
+    org.springframework.security: INFO
+    com.smartbet: INFO
+    org.apache.http: WARN
+    com.smartbet.controller: DEBUG
+    com.smartbet.service: DEBUG
+  pattern:
+    console: "%d{yyyy-MM-dd HH:mm:ss} - %msg%n"
+    file: "%d{yyyy-MM-dd HH:mm:ss} [%thread] %-5level %logger{36} - %msg%n"
+  file:
+    name: logs/smartbet-api.log
+
+management:
+  endpoints:
+    web:
+      exposure:
+        include: health,info,metrics,env
+  endpoint:
+    health:
+      show-details: always
+  info:
+    env:
+      enabled: true
+
+app:
+  thesportsdb:
+    base-url: https://www.thesportsdb.com/api/v1/json/3
+    psl-league-id: 4802
+    timeout: 30000
+    connection-timeout: 10000
+    read-timeout: 30000
+
+  football:
+    leagues:
+      psl:
+        id: 4802
+        name: "South African Premier Soccer League"
+        country: "South Africa"
+      epl:
+        id: 4328
+        name: "English Premier League"
+        country: "England"
+
+  rate-limit:
+    enabled: true
+    requests-per-minute: 60
+
+cors:
+  allowed-origins:
+    - "http://localhost:3000"
+    - "http://localhost:3001"
+    - "https://smartbet.vercel.app"
+  allowed-methods:
+    - GET
+    - POST
+    - PUT
+    - DELETE
+    - OPTIONS
+  allowed-headers: "*"
+  allow-credentials: true
